requirements.txt
Flask==0.9
Flask-Admin==1.0.2
Flask-Login==0.1.3
Flask-SQLAlchemy==0.16
Flask-Script==0.3.3
Flask-WTF==0.8
Jinja2==2.6
MySQL-python==1.2.3
SOAPpy==0.12.5
SQLAlchemy==0.7.7
WTForms==1.0.2
Werkzeug==0.8.3
argparse==1.2.1
fpconst==0.7.2
pyasn1==0.1.4
pysnmp==4.2.2
python-ldap==2.4.9
requests==1.0.3
wstools==0.4
